Committing changes 2

+++ dither.cpp
@@ -30,20 +30,20 @@ struct ppm_file_info {
 
 struct dither_settings {
     algorithm stalg;
-    double l_offset;
-    double l_scale; // -inf to +inf
-    double hue_phase;
-    double hue_scale;
-    double a_offset;
-    double b_offset;
-    double dither_intensity;
+    float l_offset;
+    float l_scale; // -inf to +inf
+    float hue_phase;
+    float hue_scale;
+    float a_offset;
+    float b_offset;
+    float dither_intensity;
 };
 
 struct palette_info {
     int num_colors;
-    std::shared_ptr<double[]> L;
-    std::shared_ptr<double[]> a;
-    std::shared_ptr<double[]> b;
+    std::shared_ptr<float[]> L;
+    std::shared_ptr<float[]> a;
+    std::shared_ptr<float[]> b;
 };
 
 void set_dither_defaults(dither_settings& settings) {
@@ -65,52 +65,52 @@ void set_dither_defaults(dither_settings& settings) {
     // --------------------------- //
 }
 
-// double -> unsigned char; clamps input between 0 and 255, then casts to unsigned char
-unsigned char cclamp(double n) {
+// float -> unsigned char; clamps input between 0 and 255, then casts to unsigned char
+unsigned char cclamp(float n) {
     if (n <= 0) return (unsigned char)0;
     if (n >= 255) return (unsigned char)255;
     return (unsigned char)std::roundf(n);
 }
 
-// double[3], double[3] -> double; euclidean distance between two 3d points
-double hypot(double a[3], double b[3]) {
-    const double x = b[0] - a[0];
-    const double y = b[1] - a[1];
-    const double z = b[2] - a[2];
+// float[3], float[3] -> float; euclidean distance between two 3d points
+float hypot(float a[3], float b[3]) {
+    const float x = b[0] - a[0];
+    const float y = b[1] - a[1];
+    const float z = b[2] - a[2];
     return x * x + y * y + z * z;
 }
 
-// double -> double; converts an srgb value to linear srgb
-double linearize(double x) {
+// float -> float; converts an srgb value to linear srgb
+float linearize(float x) {
     return powf((x + 0.055) * 0.94786729857819905, 2.4);
 }
 
-// double -> double; converts a linear srgb value to srgb
-double unlinearize(double x) {
+// float -> float; converts a linear srgb value to srgb
+float unlinearize(float x) {
     return powf(x, 0.41666666666666667) * 1.055 - 0.055;
 }
 
-// unsigned char[3] -> double[3]; converts an srgb input to oklab
-void srgb_oklab(unsigned char rgb[3], double lab[3]) {
+// unsigned char[3] -> float[3]; converts an srgb input to oklab
+void srgb_oklab(unsigned char rgb[3], float lab[3]) {
     // srgb -> linear srgb
-    const double a = linearize(double(rgb[0]) * 0.00392156862745098);
-    const double b = linearize(double(rgb[1]) * 0.00392156862745098);
-    const double c = linearize(double(rgb[2]) * 0.00392156862745098);
+    const float a = linearize(float(rgb[0]) * 0.00392156862745098);
+    const float b = linearize(float(rgb[1]) * 0.00392156862745098);
+    const float c = linearize(float(rgb[2]) * 0.00392156862745098);
     // linear srgb -> oklab
-    const double d = powf(0.4122214708 * a + 0.5363325363 * b + 0.0514459929 * c, 0.3333333333333333);
-    const double e = powf(0.2119034982 * a + 0.6806995451 * b + 0.1073969566 * c, 0.3333333333333333);
-    const double f = powf(0.0883024619 * a + 0.2817188376 * b + 0.6299787005 * c, 0.3333333333333333);
+    const float d = powf(0.4122214708 * a + 0.5363325363 * b + 0.0514459929 * c, 0.3333333333333333);
+    const float e = powf(0.2119034982 * a + 0.6806995451 * b + 0.1073969566 * c, 0.3333333333333333);
+    const float f = powf(0.0883024619 * a + 0.2817188376 * b + 0.6299787005 * c, 0.3333333333333333);
     lab[0] = 0.2104542553 * d + 0.7936177850 * e - 0.0040720468 * f;
     lab[1] = 1.9779984951 * d - 2.4285922050 * e + 0.4505937099 * f;
     lab[2] = 0.0259040371 * d + 0.7827717662 * e - 0.8086757660 * f;
 }
 
-// double[3] -> unsigned char[3]; converts an oklab input to srgb
-void oklab_srgb(double lab[3], unsigned char rgb[3]) {
+// float[3] -> unsigned char[3]; converts an oklab input to srgb
+void oklab_srgb(float lab[3], unsigned char rgb[3]) {
     // oklab -> linear srgb
-    double a = lab[0] + 0.3963377774 * lab[1] + 0.2158037573 * lab[2];
-    double b = lab[0] - 0.1055613458 * lab[1] - 0.0638541728 * lab[2];
-    double c = lab[0] - 0.0894841775 * lab[1] - 1.2914855480 * lab[2];
+    float a = lab[0] + 0.3963377774 * lab[1] + 0.2158037573 * lab[2];
+    float b = lab[0] - 0.1055613458 * lab[1] - 0.0638541728 * lab[2];
+    float c = lab[0] - 0.0894841775 * lab[1] - 1.2914855480 * lab[2];
     a = a * a * a; //std::pow(a, 3);
     b = b * b * b; //std::pow(b, 3);
     c = c * c * c; //std::pow(c, 3);
@@ -164,18 +164,18 @@ void load_palette(const std::string& palette_ppm_filename, palette_info& palette
     int num_colors = file_info.width * file_info.height;
 
     palette.num_colors = num_colors;
-    palette.L = std::shared_ptr<double[]>(new double[palette.num_colors]);
-    palette.a = std::shared_ptr<double[]>(new double[palette.num_colors]);
-    palette.b = std::shared_ptr<double[]>(new double[palette.num_colors]);
+    palette.L = std::shared_ptr<float[]>(new float[palette.num_colors]);
+    palette.a = std::shared_ptr<float[]>(new float[palette.num_colors]);
+    palette.b = std::shared_ptr<float[]>(new float[palette.num_colors]);
 
-    memset(palette.L.get(), 0, sizeof(double) * num_colors);
-    memset(palette.a.get(), 0, sizeof(double) * num_colors);
-    memset(palette.b.get(), 0, sizeof(double) * num_colors);
+    memset(palette.L.get(), 0, sizeof(float) * num_colors);
+    memset(palette.a.get(), 0, sizeof(float) * num_colors);
+    memset(palette.b.get(), 0, sizeof(float) * num_colors);
 
     unsigned palette_buffer_offset = 0;
     for (int i = 0; i < num_colors; i++) {
         unsigned char rgb[3]{};
-        double lab[3];
+        float lab[3];
 
         rgb[0] = palette_buffer[palette_buffer_offset++];
         rgb[1] = palette_buffer[palette_buffer_offset++];
@@ -196,16 +196,16 @@ void dither(const dither_settings& settings,
     int height) {
 
     int num_pixels = width * height;
-    double infinity = std::numeric_limits<double>::infinity();
+    float infinity = std::numeric_limits<float>::infinity();
 
-    const double hue_phase_cos = std::cos(settings.hue_phase * TAU);
-    const double hue_phase_sin = std::sin(settings.hue_phase * TAU);
-    const double l_scale = settings.l_scale;
-    const double hue_phase_cos_scale = hue_phase_cos * settings.hue_scale;
-    const double hue_phase_sin_scale = hue_phase_sin * settings.hue_scale;
-    const double l_offset = settings.l_offset;
-    const double a_offset = settings.a_offset;
-    const double b_offset = settings.b_offset;
+    const float hue_phase_cos = std::cos(settings.hue_phase * TAU);
+    const float hue_phase_sin = std::sin(settings.hue_phase * TAU);
+    const float l_scale = settings.l_scale;
+    const float hue_phase_cos_scale = hue_phase_cos * settings.hue_scale;
+    const float hue_phase_sin_scale = hue_phase_sin * settings.hue_scale;
+    const float l_offset = settings.l_offset;
+    const float a_offset = settings.a_offset;
+    const float b_offset = settings.b_offset;
 
     unsigned input_image_offset = 0;
     unsigned output_image_offset = 0;
@@ -215,15 +215,15 @@ void dither(const dither_settings& settings,
     int num_colors = palette.num_colors;
 
     if (settings.stalg == errordiffuse) {
-        const double negative_dither_intensity = -settings.dither_intensity;
-        double weights[ERROR_LENGTH]{};
+        const float negative_dither_intensity = -settings.dither_intensity;
+        float weights[ERROR_LENGTH]{};
         for (int i = 0; i < ERROR_LENGTH; i++) {
             weights[i] = 1.0 / (i + 1) * negative_dither_intensity;
         }
 
-        double errors_L[ERROR_LENGTH]{};
-        double errors_a[ERROR_LENGTH]{};
-        double errors_b[ERROR_LENGTH]{};
+        float errors_L[ERROR_LENGTH]{};
+        float errors_a[ERROR_LENGTH]{};
+        float errors_b[ERROR_LENGTH]{};
         for (int i = 0; i < ERROR_LENGTH; i++) {
             errors_L[i] = 0.0;
             errors_a[i] = 0.0;
@@ -231,16 +231,16 @@ void dither(const dither_settings& settings,
         }
 
         for (int i = 0; i < num_pixels; i++) {
-            double lab[3];
+            float lab[3];
             srgb_oklab(&input_image[input_image_offset], lab);
             input_image_offset += 3;
 
-            double old_lab1 = lab[1];
+            float old_lab1 = lab[1];
             lab[0] = l_offset + (lab[0] - 0.5) * l_scale + 0.5;
             lab[1] = a_offset + (lab[1] * hue_phase_cos_scale - lab[2] * hue_phase_sin_scale);
             lab[2] = b_offset + (lab[2] * hue_phase_cos_scale + old_lab1 * hue_phase_sin_scale);
 
-            double old_lab[3]{};
+            float old_lab[3]{};
             old_lab[0] = lab[0];
             old_lab[1] = lab[1];
             old_lab[2] = lab[2];
@@ -251,22 +251,22 @@ void dither(const dither_settings& settings,
             }
 
             int amin_delta = 0;
-            double best_delta = infinity;
+            float best_delta = infinity;
             for (int d = 0; d < num_colors; d++) {
 
-                double p_lab[3]{};
+                float p_lab[3]{};
                 p_lab[0] = palette_L[d];
                 p_lab[1] = palette_a[d];
                 p_lab[2] = palette_b[d];
 
-                double delta = hypot(lab, p_lab);
+                float delta = hypot(lab, p_lab);
                 if (delta < best_delta) {
                     amin_delta = d;
                     best_delta = delta;
                 }
             }
 
-            double new_lab[3]{};
+            float new_lab[3]{};
             unsigned char new_rgb[3]{};
             new_lab[0] = palette_L[amin_delta];
             new_lab[1] = palette_a[amin_delta];
@@ -293,17 +293,17 @@ void dither(const dither_settings& settings,
 
         for (int y = 0; y < height; y++) {
             for (int x = 0; x < width; x++) {
-                double lab[3];
+                float lab[3];
                 srgb_oklab(&input_image[(y * width + x) * 3], lab);
 
                 // Find the closest color in the palette to the pixel color
-                double min_dist = std::numeric_limits<double>::infinity();
+                float min_dist = std::numeric_limits<float>::infinity();
                 int min_idx = -1;
                 for (int i = 0; i < num_colors; i++) {
-                    double L_diff = lab[0] - palette_L[i];
-                    double a_diff = lab[1] - palette_a[i];
-                    double b_diff = lab[2] - palette_b[i];
-                    double dist = std::sqrt(L_diff * L_diff + a_diff * a_diff + b_diff * b_diff);
+                    float L_diff = lab[0] - palette_L[i];
+                    float a_diff = lab[1] - palette_a[i];
+                    float b_diff = lab[2] - palette_b[i];
+                    float dist = std::sqrt(L_diff * L_diff + a_diff * a_diff + b_diff * b_diff);
                     if (dist < min_dist) {
                         min_dist = dist;
                         min_idx = i;
@@ -313,8 +313,8 @@ void dither(const dither_settings& settings,
                 // Dither using the closest color in the palette
                 int ix = x % matrix_size;
                 int iy = y % matrix_size;
-                double threshold = (matrix[iy][ix] + 0.5) / (matrix_size * matrix_size);
-                double l = lab[0];
+                float threshold = (matrix[iy][ix] + 0.5) / (matrix_size * matrix_size);
+                float l = lab[0];
                 if (l <= threshold) {
                     l = palette_L[min_idx];
                 }
